+++ braintree/braintree_gateway.py
@@ -6,6 +6,7 @@ from braintree.credit_card_gateway import CreditCardGateway
 from braintree.credit_card_verification_gateway import CreditCardVerificationGateway
 from braintree.customer_gateway import CustomerGateway
 from braintree.discount_gateway import DiscountGateway
+from braintree.dispute_gateway import DisputeGateway
 from braintree.merchant_account_gateway import MerchantAccountGateway
 from braintree.merchant_gateway import MerchantGateway
 from braintree.oauth_gateway import OAuthGateway
@@ -41,6 +42,7 @@ class BraintreeGateway(object):
         self.credit_card = CreditCardGateway(self)
         self.customer = CustomerGateway(self)
         self.discount = DiscountGateway(self)
+        self.dispute = DisputeGateway(self)
         self.merchant_account = MerchantAccountGateway(self)
         self.merchant = MerchantGateway(self)
         self.oauth = OAuthGateway(self)