+++ braintree/dispute.py
@@ -1,6 +1,8 @@
 from decimal import Decimal
 from braintree.attribute_getter import AttributeGetter
 from braintree.transaction_details import TransactionDetails
+from braintree.dispute_details import DisputeEvidence, DisputeStatusHistory
+from braintree.configuration import Configuration
 
 class Dispute(AttributeGetter):
     class Status(object):
@@ -55,11 +57,23 @@ class Dispute(AttributeGetter):
         PreArbitration = "pre_arbitration"
         Retrieval      = "retrieval"
 
+    @staticmethod
+    def find(id):
+        return Configuration.gateway().dispute.find(id)
 
     def __init__(self, attributes):
         AttributeGetter.__init__(self, attributes)
 
-        if self.amount is not None:
+        if "amount" in attributes and self.amount is not None:
             self.amount = Decimal(self.amount)
+        if "amount_disputed" in attributes and self.amount_disputed is not None:
+            self.amount_disputed = Decimal(self.amount_disputed)
+        if "amount_won" in attributes and self.amount_won is not None:
+            self.amount_won = Decimal(self.amount_won)
         if "transaction" in attributes:
             self.transaction_details = TransactionDetails(attributes.pop("transaction"))
+            self.transaction = self.transaction_details
+        if "evidence" in attributes and self.evidence is not None:
+            self.evidence = [DisputeEvidence(evidence) for evidence in self.evidence]
+        if "status_history" in attributes and self.status_history is not None:
+            self.status_history = [DisputeStatusHistory(status_history) for status_history in self.status_history]